Do not log info messages when verbosity is quiet
Depending on the module and exact situation config settings can contain
sensitive information which is outputted to the console. As this is an
info message there is no way to supress this by default.

Only log info messages when verbosity is not set to quiet to allow
suppression

(cherry picked from commit 126c32313838988a527d4980384a561c20fd31f0)